Collab backend : Fix bug which caused E11000 duplicate key error
Problem is caused when both users enter a new room, both check that room does not have an associated file, and both try to make one.
This led to the a the error, when the 2nd user tried to add a new entry with the same id.
The solution is to use mutex to lock it. An alternate solution would be to use a Message Broker/Queue, but this seemed to be simpler